refactor(shell): split native host runtime ownership
