+++ packages/shell-host/lib/installer.d.mts
@@ -30,3 +30,9 @@ export function createNativeHostManifest(
   args: Pick<InstallerArgs, 'browser' | 'extensionId'>,
   wrapperPath: string,
 ): Record<string, unknown>;
+export function copyHostRuntime(nativeSourceDir: string, installDir: string): string;
+export function getMissingHostRuntimeFiles(installDir: string): string[];
+export function createWrapper(hostPath: string, logFile: string | null): string;
+export function detectLogFile(wrapperPath: string): string | null;
+export function escapeShellValue(value: string): string;
+export function main(argv?: string[]): Promise<void>;
